+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Unreleased
 - Added ActionResult.
 - Added type hints functionality in action method parameters.
+- Added ActionResultException handling in action methods.
 - Deprecated AbstractActionResult.
 
 ## [2.0.1] 2018-06-19